profileservice/requirements.txt
django
djangorestframework
django-cors-headers
pymongo